fix: unable to open update dialog on preference page
