+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - 前提アセット選択欄がダイアログからはみ出ることがある問題を修正
 - ドラッグアンドドロップでファイルを追加できなくなることがある問題を修正
 - ZIP展開時のプログレスバーの挙動を改善
+- 設定画面でアップデートダイアログが開けないバグを修正
 
 ### ⚙️その他
