+++ src/components/page/Top/hook.tsx
@@ -1,22 +1,13 @@
 import { AssetContextType } from '@/components/context/AssetContext'
 import { useState, useEffect, useContext } from 'react'
-import {
-  checkForUpdate,
-  dismissUpdate,
-  downloadUpdate,
-  onFileDrop,
-  refreshAssets,
-} from './logic'
-import { useToast } from '@/hooks/use-toast'
-import { buttonVariants } from '@/components/ui/button'
-import { ToastAction } from '@/components/ui/toast'
+import { onFileDrop, refreshAssets } from './logic'
 import { PersistentContext } from '@/components/context/PersistentContext'
 import { AssetSummary } from '@/lib/bindings'
-import { useLocalization } from '@/hooks/use-localization'
 import {
   DragDropContext,
   DragDropRegisterConfig,
 } from '@/components/context/DragDropContext'
+import { UpdateDialogContext } from '@/components/context/UpdateDialogContext'
 
 type ReturnProps = {
   assetContextValue: AssetContextType
@@ -29,13 +20,10 @@ type ReturnProps = {
   setEditAssetDialogOpen: (open: boolean) => void
   editAssetDialogAssetId: string | null
   editAssetDialogOpen: boolean
-  updateDialogOpen: boolean
-  setUpdateDialogOpen: (open: boolean) => void
   dataManagementDialogAssetId: string | null
   dataManagementDialogOpen: boolean
   setDataManagementDialogOpen: (open: boolean) => void
   openDataManagementDialog: (assetId: string) => void
-  updateDownloadTaskId: string | null
 }
 
 export const useTopPage = (): ReturnProps => {
@@ -50,11 +38,6 @@ export const useTopPage = (): ReturnProps => {
     string | null
   >(null)
 
-  const [updateDialogOpen, setUpdateDialogOpen] = useState(false)
-  const [updateDownloadTaskId, setUpdateDownloadTaskId] = useState<
-    string | null
-  >(null)
-
   const [dataManagementDialogOpen, setDataManagementDialogOpen] =
     useState(false)
   const [dataManagementDialogAssetId, setDataManagementDialogAssetId] =
@@ -64,10 +47,8 @@ export const useTopPage = (): ReturnProps => {
 
   const [isDragAndHover, setDragAndHover] = useState(false)
 
-  const { t } = useLocalization()
-  const { toast } = useToast()
-
   const { register } = useContext(DragDropContext)
+  const { checkForUpdate } = useContext(UpdateDialogContext)
 
   const assetContextValue: AssetContextType = {
     assetDisplaySortedList: assetDisplaySortedList,
@@ -100,44 +81,8 @@ export const useTopPage = (): ReturnProps => {
     })
   }, [])
 
-  const startUpdateDownloadingAndOpenDialog = async () => {
-    await downloadUpdate(setUpdateDownloadTaskId)
-    setUpdateDialogOpen(true)
-  }
-
   const executeUpdateCheck = async () => {
-    const updateAvailable = await checkForUpdate()
-
-    if (!updateAvailable) {
-      return
-    }
-
-    toast({
-      title: t('top:new-update-toast'),
-      description: t('top:new-update-toast:description'),
-      duration: 30000,
-      onSwipeEnd: () => {
-        dismissUpdate()
-      },
-      action: (
-        <div className="flex flex-col space-y-2">
-          <ToastAction
-            altText="Open update dialog"
-            className={buttonVariants({ variant: 'default' })}
-            onClick={startUpdateDownloadingAndOpenDialog}
-          >
-            {t('top:new-update-toast:button')}
-          </ToastAction>
-          <ToastAction
-            altText="Ignore update"
-            className={buttonVariants({ variant: 'outline' })}
-            onClick={() => dismissUpdate()}
-          >
-            {t('top:new-update-toast:close')}
-          </ToastAction>
-        </div>
-      ),
-    })
+    await checkForUpdate()
   }
 
   useEffect(() => {
@@ -160,12 +105,9 @@ export const useTopPage = (): ReturnProps => {
     setEditAssetDialogOpen,
     editAssetDialogAssetId,
     editAssetDialogOpen,
-    updateDialogOpen,
-    setUpdateDialogOpen,
     dataManagementDialogAssetId,
     dataManagementDialogOpen,
     setDataManagementDialogOpen,
     openDataManagementDialog,
-    updateDownloadTaskId,
   }
 }
